♻️ refactor: storiesLayout 분리
레이아웃과 PageContent 분리로 테스트 용이성 증대
